requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
beautifulsoup4==4.8.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.7
ffmpeg-python==0.2.0
future==0.17.1
gTTS==2.0.4
gTTS-token==1.1.3
gunicorn==19.9.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycparser==2.19
pylint==2.3.1
python-telegram-bot==12.1.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
SpeechRecognition==3.8.1
tornado==6.0.3
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
